+++ src/main/java/janggi/domain/board/PalaceArea.java
@@ -1,4 +1,4 @@
-package janggi.board;
+package janggi.domain.board;
 
 import java.util.Set;
