+++ src/main/java/janggi/domain/piece/type/MoveType.java
@@ -1,4 +1,4 @@
-package janggi.piece;
+package janggi.domain.piece.type;
 
 public enum MoveType {
     NORMAL, PALACE;
