+++ src/main/java/janggi/domain/piece/type/PieceType.java
@@ -1,6 +1,14 @@
-package janggi.piece;
+package janggi.domain.piece.type;
 
-import janggi.camp.Camp;
+import janggi.domain.camp.Camp;
+import janggi.domain.piece.Cannon;
+import janggi.domain.piece.Chariot;
+import janggi.domain.piece.Elephant;
+import janggi.domain.piece.General;
+import janggi.domain.piece.Guard;
+import janggi.domain.piece.Horse;
+import janggi.domain.piece.Piece;
+import janggi.domain.piece.Soldier;
 
 public enum PieceType {
